requirements.txt
torch==2.1.0
transformers==4.36.0
sentence-transformers==2.3.1
langchain==0.1.0
langchain-community==0.0.10
openai==1.6.1
faiss-cpu==1.7.4
pandas==2.1.4
numpy==1.26.2
mlflow==2.9.2
fastapi
uvicorn
pydantic
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.2
tiktoken==0.5.2
chromadb==0.4.22
prophet==1.1.5
plotly==5.18.0
streamlit==1.29.0
watchdog==3.0.0